pyproject.toml
[project]
name = "pythemonsters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ascii-magic>=2.3.0",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
]
